+++ Example.hs
@@ -34,20 +34,23 @@ help1 args = liftIO $ print $ "Help!" ++ show args
 puts1 :: [String] -> Repl1 ()
 puts1 args = modify $ Set.union (Set.fromList args)
 
-opts1 :: [(String, [String] -> Repl1 ())]
+opts1 :: [(String, String -> Repl1 ())]
 opts1 =
-  [ ("help", help1), -- :help
-    ("puts", puts1) -- :puts
+  [ ("help", help1 . words), -- :help
+    ("puts", puts1 . words) -- :puts
   ]
 
 init1 :: Repl1 ()
 init1 = return ()
 
+final1 :: Repl1 ExitDecision
+final1 = return Exit
+
 -- Tab completion inside of StateT
 repl1 :: IO ()
 repl1 =
   flip evalStateT Set.empty $
-    evalRepl (pure "_proto> ") cmd1 opts1 (Just ':') (Word completer1) init1
+    evalRepl (const $ pure "_proto> ") cmd1 opts1 (Just ':') Nothing (Word completer1) init1 final1
 
 -------------------------------------------------------------------------------
 -- Command options
@@ -67,16 +70,21 @@ comp2 = listWordCompleter ["kirk", "spock", "mccoy"]
 help2 :: [String] -> Repl2 ()
 help2 args = liftIO $ print $ "Help!" ++ show args
 
-opts2 :: [(String, [String] -> Repl2 ())]
+opts2 :: [(String, String -> Repl2 ())]
 opts2 =
-  [ ("help", help2)
+  [ ("help", help2 . words)
   ]
 
 init2 :: Repl2 ()
 init2 = liftIO $ putStrLn "Welcome!"
 
+final2 :: Repl2 ExitDecision
+final2 = do
+  liftIO $ putStrLn "Goodbye!"
+  return Exit
+
 repl2 :: IO ()
-repl2 = evalRepl (pure "example2> ") cmd2 opts2 (Just ':') (Word comp2) init2
+repl2 = evalRepl (const $ pure "example2> ") cmd2 opts2 (Just ':') Nothing (Word comp2) init2 final2
 
 -------------------------------------------------------------------------------
 -- Mixed Completion
@@ -99,17 +107,17 @@ byWord n = do
   let names = ["picard", "riker", "data", ":file", ":holiday"]
   return $ filter (isPrefixOf n) names
 
-files :: [String] -> Repl3 ()
+files :: String -> Repl3 ()
 files args = liftIO $ do
-  contents <- readFile (unwords args)
+  contents <- readFile args
   putStrLn contents
 
-holidays :: [String] -> Repl3 ()
-holidays [] = liftIO $ putStrLn "Enter a holiday."
+holidays :: String -> Repl3 ()
+holidays "" = liftIO $ putStrLn "Enter a holiday."
 holidays xs = liftIO $ do
-  putStrLn $ "Happy " ++ unwords xs ++ "!"
+  putStrLn $ "Happy " ++ xs ++ "!"
 
-opts3 :: [(String, [String] -> Repl3 ())]
+opts3 :: [(String, String -> Repl3 ())]
 opts3 =
   [ ("file", files),
     ("holiday", holidays)
@@ -118,8 +126,11 @@ opts3 =
 init3 :: Repl3 ()
 init3 = return ()
 
+final3 :: Repl3 ExitDecision
+final3 = return Exit
+
 repl3 :: IO ()
-repl3 = evalRepl (pure "example3> ") cmd3 opts3 (Just ':') (Prefix (wordCompleter byWord) defaultMatcher) init3
+repl3 = evalRepl (const $ pure "example3> ") cmd3 opts3 (Just ':') Nothing (Prefix (wordCompleter byWord) defaultMatcher) init3 final3
 
 -------------------------------------------------------------------------------
 --