+++ examples/Simple.hs
@@ -21,32 +21,43 @@ completer n = do
 help :: [String] -> Repl ()
 help args = liftIO $ print $ "Help: " ++ show args
 
-say :: [String] -> Repl ()
+say :: String -> Repl ()
 say args = do
-  _ <- liftIO $ callCommand $ "cowsay" ++ " " ++ (unwords args)
+  _ <- liftIO $ callCommand $ "cowsay" ++ " " ++ args
   return ()
 
-opts :: [(String, [String] -> Repl ())]
+opts :: [(String, String -> Repl ())]
 opts =
-  [ ("help", help), -- :help
+  [ ("help", help . words), -- :help
     ("say", say) -- :say
   ]
 
 ini :: Repl ()
 ini = liftIO $ putStrLn "Welcome!"
 
+final :: Repl ExitDecision
+final = do
+  liftIO $ putStrLn "Goodbye!"
+  return Exit
+
 repl_alt :: IO ()
 repl_alt = evalReplOpts $ ReplOpts
-  { banner      = pure ">>> "
-  , command     = cmd
-  , options     = opts
-  , prefix      = Just ':'
-  , tabComplete = (Word0 completer)
-  , initialiser = ini
+  { banner           = const $ pure ">>> "
+  , command          = cmd
+  , options          = opts
+  , prefix           = Just ':'
+  , multilineCommand = Just "paste"
+  , tabComplete      = (Word0 completer)
+  , initialiser      = ini
+  , finaliser        = final
   }
 
+customBanner :: MultiLine -> Repl String
+customBanner SingleLine = pure ">>> "
+customBanner MultiLine = pure "| "
+
 repl :: IO ()
-repl = evalRepl (pure ">>> ") cmd opts (Just ':') (Word0 completer) ini
+repl = evalRepl (const $ pure ">>> ") cmd opts (Just ':') (Just "paste") (Word0 completer) ini final
 
 main :: IO ()
 main = pure ()