+++ examples/Stateful.hs
@@ -13,18 +13,18 @@ import System.Console.Repline
 -- Stateful Completion
 -------------------------------------------------------------------------------
 
-type IState = Set.Set String
+type IState = (Int, Set.Set String)
 
 type Repl a = HaskelineT (StateT IState IO) a
 
 -- Evaluation
 cmd :: String -> Repl ()
-cmd input = modify $ \s -> Set.insert input s
+cmd input = modify . fmap $ \s -> Set.insert input s
 
 -- Completion
 comp :: (Monad m, MonadState IState m) => WordCompleter m
 comp n = do
-  ns <- get
+  (c, ns) <- get
   return $ filter (isPrefixOf n) (Set.toList ns)
 
 -- Commands
@@ -32,22 +32,32 @@ help :: [String] -> Repl ()
 help args = liftIO $ print $ "Help!" ++ show args
 
 puts :: [String] -> Repl ()
-puts args = modify $ \s -> Set.union s (Set.fromList args)
+puts args = modify . fmap $ \s -> Set.union s (Set.fromList args)
 
-opts :: [(String, [String] -> Repl ())]
+opts :: [(String, String -> Repl ())]
 opts =
-  [ ("help", help), -- :help
-    ("puts", puts) -- :puts
+  [ ("help", help . words), -- :help
+    ("puts", puts . words) -- :puts
   ]
 
 ini :: Repl ()
 ini = return ()
 
+final :: Repl ExitDecision
+final = do
+  (count, s) <- get
+  if count == 0
+  then return Exit
+  else do
+    liftIO . putStrLn $ "Exit in " <> show count <> "..."
+    put (count - 1, s)
+    return Continue
+
 -- Tab completion inside of StateT
 repl :: IO ()
 repl =
-  flip evalStateT Set.empty $
-    evalRepl (pure ">>> ") cmd opts Nothing (Word comp) ini
+  flip evalStateT (3, Set.empty) $
+    evalRepl (const $ pure ">>> ") cmd opts Nothing Nothing (Word comp) ini final
 
 main :: IO ()
 main = pure ()